refactor: separate handlers, add getResponse func, rename variables, clear types
